Refactor PeerController and PeerHanle into a seperate file
